+++ jvm/src/main/scala/eu/cdevreeze/yaidom/dom/domNode.scala
@@ -30,7 +30,6 @@ import eu.cdevreeze.yaidom.queryapi.HasParent
 import eu.cdevreeze.yaidom.queryapi.Nodes
 import eu.cdevreeze.yaidom.queryapi.ScopedElemNodeApi
 import eu.cdevreeze.yaidom.queryapi.ScopedElemLike
-import eu.cdevreeze.yaidom.resolved.ResolvedNodes
 
 /**
  * Wrappers around `org.w3c.dom.Node` and subclasses, such that the wrapper around `org.w3c.dom.Element` conforms to the
@@ -47,7 +46,7 @@ import eu.cdevreeze.yaidom.resolved.ResolvedNodes
  *
  * @author Chris de Vreeze
  */
-sealed trait DomNode extends ResolvedNodes.Node {
+sealed trait DomNode extends Nodes.Node {
 
   type DomType <: w3c.dom.Node
 
@@ -83,7 +82,7 @@ sealed trait CanBeDomDocumentChild extends DomNode with Nodes.CanBeDocumentChild
  * this DomElem makes namespace-aware querying of DOM elements far easier than direct querying of DOM elements.
  */
 final class DomElem(
-  override val wrappedNode: w3c.dom.Element) extends CanBeDomDocumentChild with ResolvedNodes.Elem with ScopedElemNodeApi with ScopedElemLike with HasParent {
+  override val wrappedNode: w3c.dom.Element) extends CanBeDomDocumentChild with Nodes.Elem with ScopedElemNodeApi with ScopedElemLike with HasParent {
 
   require(wrappedNode ne null) // scalastyle:off null
 
@@ -170,7 +169,7 @@ final class DomElem(
   }
 }
 
-final class DomText(override val wrappedNode: w3c.dom.Text) extends DomNode with ResolvedNodes.Text {
+final class DomText(override val wrappedNode: w3c.dom.Text) extends DomNode with Nodes.Text {
   require(wrappedNode ne null) // scalastyle:off null
 
   override type DomType = w3c.dom.Text