+++ jvm/src/test/scala/eu/cdevreeze/yaidom/integrationtest/XmlDiffTest.scala
@@ -25,7 +25,8 @@ import eu.cdevreeze.yaidom.core.Path
 import eu.cdevreeze.yaidom.core.QName
 import eu.cdevreeze.yaidom.indexed
 import eu.cdevreeze.yaidom.parse.DocumentParserUsingDom
-import eu.cdevreeze.yaidom.queryapi.ClarkElemApi
+import eu.cdevreeze.yaidom.queryapi.ClarkElemNodeApi
+import eu.cdevreeze.yaidom.queryapi.Nodes
 import eu.cdevreeze.yaidom.resolved
 import eu.cdevreeze.yaidom.simple
 
@@ -144,7 +145,7 @@ object XmlDiffTest {
    * Both issues are resolvable. The matching function could be made pluggable. Using IndexedClarkElems is still
    * useful, for difference reporting purposes.
    */
-  def findDiffs[E <: resolved.ResolvedNodes.Elem with ClarkElemApi.Aux[E]](elem1: E, elem2: E): Diffs = {
+  def findDiffs[E <: Nodes.Elem with ClarkElemNodeApi.Aux[_, E]](elem1: E, elem2: E): Diffs = {
     val indexedResolvedElem1 =
       indexed.IndexedClarkElem(resolved.Elem(elem1).removeAllInterElementWhitespace.coalesceAndNormalizeAllText)
     val indexedResolvedElem2 =
@@ -174,9 +175,9 @@ object XmlDiffTest {
   }
 
   final case class Diffs(
-      val onlyInFirstElem: Set[Path],
-      val onlyInSecondElem: Set[Path],
-      val inBothElemsButDiffering: Set[Path]) {
+    val onlyInFirstElem:         Set[Path],
+    val onlyInSecondElem:        Set[Path],
+    val inBothElemsButDiffering: Set[Path]) {
 
     def allDiffs = onlyInFirstElem.union(onlyInSecondElem).union(inBothElemsButDiffering)
   }