+++ shared/src/main/scala/eu/cdevreeze/yaidom/simple/node.scala
@@ -33,7 +33,6 @@ import eu.cdevreeze.yaidom.queryapi.ScopedElemNodeApi
 import eu.cdevreeze.yaidom.queryapi.ScopedElemLike
 import eu.cdevreeze.yaidom.queryapi.TransformableElemLike
 import eu.cdevreeze.yaidom.queryapi.UpdatableElemLike
-import eu.cdevreeze.yaidom.resolved.ResolvedNodes
 
 /**
  * Immutable XML Node. It is the default XML node type in yaidom. There are subclasses for different types of nodes,
@@ -42,7 +41,7 @@ import eu.cdevreeze.yaidom.resolved.ResolvedNodes
  *
  * @author Chris de Vreeze
  */
-sealed trait Node extends ResolvedNodes.Node with Immutable with Serializable {
+sealed trait Node extends Nodes.Node with Immutable with Serializable {
 
   /**
    * Returns the tree representation String, conforming to the tree representation DSL that creates `NodeBuilder`s.
@@ -176,7 +175,7 @@ final class Elem(
   val scope:             Scope,
   override val children: immutable.IndexedSeq[Node])
   extends CanBeDocumentChild
-  with ResolvedNodes.Elem
+  with Nodes.Elem
   with ScopedElemNodeApi
   with ScopedElemLike
   with UpdatableElemLike
@@ -678,7 +677,7 @@ final class Elem(
 }
 
 @SerialVersionUID(1L)
-final case class Text(text: String, isCData: Boolean) extends Node with ResolvedNodes.Text {
+final case class Text(text: String, isCData: Boolean) extends Node with Nodes.Text {
   require(text ne null) // scalastyle:off null
   if (isCData) require(!text.containsSlice("]]>"))